Fix the PartialEq impl of Field. (#30)
It is currently assumed that `self` in the `PartialEq::eq`
implementation of `Field` always has a magnitude of 1,
otherwise triggering an assertion in debug mode or (wrongly)
operating with `m = 1` in release mode within `Field::neg_in_place`
in the context of comparing public keys. In practice, that means
equality of public keys can fail to be symmetric.

In particular, the magnitude of the `y` coordinate of a public key
can be `2` after serializing and deserializing it in compressed form,
since if the `y` bit is odd, there is another negation in
`Affine::set_xo_var` (called from `PublicKey::parse_compressed`)
which increases the magnitude.

This is addressed here by using the current (assumed to be valid)
`magnitude` of the field element being negated as the maximum
magnitude in `Field::neg_in_place`. The reasoning being that
the `PartialOrd` impl should work for elements with any
magnitude.